requirements.txt
spotipy
pandas
numpy
scikit-learn
python-dotenv
joblib